p2-c3-sc1: Double multiplication with BigDecimal

+++ src/main/java/com/openclassrooms/testing/calculator/service/BatchCalculator.java
@@ -48,7 +48,7 @@ public class BatchCalculator {
     private CalculationModel solve(CalculationModel calculationModel) {
         CalculationType type = calculationModel.getType();
 
-        Integer response = null;
+        Double response = null;
         switch (type) {
             case ADDITION:
                 response = calculator.add(
