+++ src/test/java/com/openclassrooms/testing/calculator/service/BatchCalculatorComponentIntegrationTest.java
@@ -69,40 +69,40 @@ public class BatchCalculatorComponentIntegrationTest {
     public void calculateFromFile_shouldReturnTheCorrectAnswer_forAdditions() throws IOException {
         // ACT
         List<CalculationModel> solutions = classUnderTest.calculateFromFile(FIXTURE_FILE);
-        Integer answer = solutions.get(ADDITION_INDEX).getSolution();
+        Double answer = solutions.get(ADDITION_INDEX).getSolution();
 
         // ASSERT we get back usable models
-        assertThat(answer, is(equalTo(4)));
+        assertThat(answer, is(equalTo(4.0)));
     }
 
     @Test
     public void calculateFromFile_shouldCorrectlyAddWithTheCalculator_forAdditions() throws IOException {
         // ACT
         List<CalculationModel> solutions = classUnderTest.calculateFromFile(FIXTURE_FILE);
-        Integer answer = solutions.get(ADDITION_INDEX).getSolution();
+        Double answer = solutions.get(ADDITION_INDEX).getSolution();
 
         // ASSERT we get back usable models
-        verify(calculator, times(1)).add(2, 2);
+        verify(calculator, times(1)).add(2.0, 2.0);
     }
 
     @Test
     public void calculateFromFile_shouldReturnTheCorrectAnswer_forMultiplication() throws IOException {
         // ACT
         List<CalculationModel> solutions = classUnderTest.calculateFromFile(FIXTURE_FILE);
-        Integer answer = solutions.get(MULTIPLICATION_INDEX).getSolution();
+        Double answer = solutions.get(MULTIPLICATION_INDEX).getSolution();
 
         // ASSERT we get back usable models
-        assertThat(answer, is(equalTo(6)));
+        assertThat(answer, is(equalTo(6.0)));
     }
 
     @Test
     public void calculateFromFile_shouldCorrectlyMultiplyWithTheCalculator_forProducts() throws IOException {
         // ACT
         List<CalculationModel> solutions = classUnderTest.calculateFromFile(FIXTURE_FILE);
-        Integer answer = solutions.get(ADDITION_INDEX).getSolution();
+        Double answer = solutions.get(ADDITION_INDEX).getSolution();
 
         // ASSERT we get back usable models
-        verify(calculator, times(1)).multiply(3, 2);
+        verify(calculator, times(1)).multiply(3.0, 2.0);
     }
 
     @Test
@@ -113,10 +113,10 @@ public class BatchCalculatorComponentIntegrationTest {
 
         // ASSERT we get back usable models
         assertThat(answer, allOf(
-                hasProperty("leftArgument", is(2)),
-                hasProperty("rightArgument", is(2)),
+                hasProperty("leftArgument", is(2.0)),
+                hasProperty("rightArgument", is(2.0)),
                 hasProperty("type", is(equalTo(CalculationType.ADDITION))),
-                hasProperty("solution", is(4))));
+                hasProperty("solution", is(4.0))));
     }
